feat: add systemd socket activation support to the master
This change allows the master to be run using systemd's socket
activation feature. It simply changes the creation of the base socket
listener to check for one provided by systemd or fall back to a manually
created listener; everything else built on that stays the same.

We also ship a new socket unit file so that users can switch to socket
activation easily with a few `systemctl` invocations.